Support scalafixDependencies
This PR improves support for custom rules that are published to Maven
Central. This functionality is required for rules like
scala-collection-compat. We do a best-effort to avoid redundant
dependency fetching of `scalafixDependencies` by caching the results
in an in-memory map. In the future, it would make sense to expose more
powerful APIs in scalafix to cache classloaders between invocations.