Upgrade to scalafix M14
This upgrade introduced a new Java API in a module scalafix-interfaces
that exposes a public API for build tools and IDEs
https://github.com/scalacenter/scalafix/pull/783. This means
sbt-scalafix no longer needs to invoke the fragile
`main(args: Array[String]): Unit` interface. This also means it's
possible to hook into the linter reporting pipeline and use the
sbt logger for diagnostics.