Explicitly pass around configuration for config settings
Attempt to fix https://github.com/scalacenter/scalafix/issues/692.
Previously, the scalafix command invoked an unscoped `compile` inside a
dynamic task. Now, the plugin scopes the command to the configuration.
This commit adds a scripted tests to test configurations in isolation
and it seems to work according to the expectations in the linked ticket:

>  invoking it:scalafixCli or resilience:scalafixCli should only run
>  scalafix for that test configuration

Happy to iterate further if necessary.