Migrate from coursier-small to coursier:interface
The coursier-small module was created before the official
coursier/interface module was available. They both have the same goal,
expose a stable, zero dependency, API for fetching dependencies.  The
official API has the benefit of playing nicely with the official
coursier libraries when running concurrently.

The coursier/interface API is also richer, it for example supports
credentials, which is necessary to fix
https://github.com/scalacenter/scalafix/issues/984

We can't fully get rid of coursier-small yet since we use its types  in
the public API of sbt-scalafix. We should ideally deprecate the
coursier-small based settings down the road.