deprecate usage of coursiersmall type in scalafixResolvers
This is breaking binary compatibility, but preserves source
compatibility via (deprecate) implicit conversion, which is always
in scope for *.sbt sources, but require an explicit import for *.scala
sources, such as `import scalafix.sbt.ScalafixPlugin.autoImport._`.

That means that most 99% of direct sbt-scalafix users should be fine,
but any plugin published against sbt-scalafix will need to be rebuilt.

Usage of coursierapi is a bit more cumbersome for sbt 0.13.x because
the Java8 static methods in Repository cannot be used. As the scripted
shows, it's still possible to use it without these helpers.