delegate rules fetching and classloading to scalafix-interfaces
The class loading hierarchy effectively becomes one level flatter,
as the potential (5) has now (3) as parent, which unfortunately results
in a small regression from 568f1a2 in case `refreshClasspath` is
true, which happens for invocations that either:
 (1) request external rules via CLI (`dependency:`)
 (2) use local rules (Scalafix ivy config)

In this case the potential `scalafixDependencies` are unnecessarily
fetched & reloaded each time (see `scalafixArgsFromShell()`).

If needed, we can later memoize the state of the `scalafixArguments`
instance containing that classloader across invocations if only (2)
is involved, relying on stamping local rules bytecode to detect when
invalidation is needed.

For (1), the performance penalty is not a big deal anyway as this is
probably meant to be a one-shot run, as the user with some
`scalafixDependencies` will probably move away that CLI-syntax if
it is meant to be permanent.