introduce scalafixOnCompile
Users can control whether scalafix (running the default rules declared
in .scalafix.conf) should be run as part of `compile` at the project &
configuration level.

The value of scalafixOnCompile is ignored when invoking directly
scalafix or scalafixAll, as the CLI arguments (if there are any)
take precedence over the default rules. That means for example that
`scalafix --check` is safe to run even though scalafixOnCompile := true,
as there will not be any rewrite triggered by the implicit compilation.