share custom toolclasspath classloader across projects & configs
Since d81dd72 (see commit description), when `scalafix` or `scalafixAll`
is run on a build root that aggregates many projects, many identical
classloaders may be created when an external rule is provided on the
CLI or when projects use local rules via ScalafixConfig.

This allows aggregated tasks across projects & configs to share a
single, warm classloader when possible (same Scalafix classpath).

Note that cache is bound on purpose to a task value, so two successive
invocations of `scalafix` with the exact same arguments & environment
will not share the same classloader. This is for simplicity, as local
rules can be updated from one run to the other, and to my knowledge,
there is no simple way to invalidate a URLClassLoader if the underlying
class files change.