caching: remove optimization bringing false positive cache hits
If a file is patched during a scalafix run, it's not safe to assume
that all rules observed the updated content of that file and thus can
be skipped next time. It is if we have only one rule (which we could
check) or if the patch is applied after the first rule (which we can't
tell). For correctness, stamping should happen before rule execution.