caching: don't forget previous invocations
This allows cache accumulation after successive runs of scalafix
with different rules, which is common as it's currently the only
way to safely run several semantic rules in order.

Note that the cache remains incremental only for files, so if a file
is stale for a single rule, all rules will be executed on this file.