prevent concurrent modifications of sources by limiting concurrency
When rewrite rules are executed, `scalafix` and `scalafixAll` may
conflict with other tasks transforming `unmanagedSources`. Since
there is no way to identify such tasks given the sbt model, and to
be defensive against this scenario, `scalafix` tasks may only run
concurrently with other `scalafix` tasks, but not with any other
tasks.

For explicit invocations, this should not damage throughput, but for
complex builds with `scalafixOnCompile := true`, this can have a
limited impact during the first compile; after that caching should
make it negligible.