scalafixEnable: relax scalacOptions only when no other task is requested
Even though the built-in `all` command has effectively never supported
input tasks (https://github.com/sbt/sbt/commit/21a0597), a custom command
or a future iteration of `all` could could, making it possible to run
an input task in parallel with other tasks or input tasks. In that case,
scalacOptions should remain unchanged as we might for example end up
packaging classes that the compiler raised warnings for (if `package` is
the other task scheduled alongside scalafix) despite the -Xfatal-warnings
requested in the build.