scalafixEnable: scalafix should not trigger compilation when not needed
IncOptions#withIgnoredScalacOptions allows to temporarily leverage the
cache when the last build was successful, despite having different
scalacOptions. It is safe when scalafix is invoked, as the updated
options are more permissive, but do not change the outcome of the class
files nor the semanticdb files.

The Zinc analysis file is not saved when compilation is triggered
through scalafix as the task depends on manipulateBytecode which does
not save the report.