scalafixEnable: *latest* semanticdb for old scala
When there is no semanticdb-scalac available for the build scala version for
the scalameta version that scalafix tracks, we can tell the user if it is
because it is no longer published or not yet published. When it is no longer
published, the latest version should not change, so we can use that one in a
repeatable manner, in order to maximize the chances of compatiblity.