dirty hack to support non-canonical scalameta version tag
On sbt 1.4.0, this prevents "java.lang.IllegalArgumentException: Invalid
comparator: >4.12.4.1".

On recent sbt versions, "4.12.4.1" is supported, but not parsed the way one
would expect, so it's safer to take out the last digit explicitly in all
cases, since the intent of that pattern was to indicate that this should be
used over 4.12.4.

https://github.com/sbt/librarymanagement/pull/378